feat: solve cjsToEsm.cjs task
